+++ app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -1,0 +1,9 @@
+import { SignUp } from "@clerk/nextjs";
+
+export default function Page() {
+  return (
+    <div className="h-screen flex justify-center items-center">
+      <SignUp />;
+    </div>
+  );
+}
