+++ app/(dashboard)/page.tsx
@@ -1,0 +1,9 @@
+import { UserButton } from "@clerk/nextjs";
+
+export default function Home() {
+  return (
+    <div>
+      <UserButton />
+    </div>
+  );
+}
